# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Use auto API documentation",
        "contact": {
            "email": "and45@mail.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.1"
    },
    "servers": [
        {
            "url": "http://autouse/api/V1",
            "description": "Use auto API server"
        }
    ],
    "paths": {
        "/cars": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Display a listing of the resource",
                "operationId": "carsAll",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarShowRequest"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "Create yet another example record",
                "operationId": "Create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarBusyStore"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarShowRequest"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/cars/{id}": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Get car by ID",
                "description": "Display the specified resource.",
                "operationId": "carGet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of car",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarShowRequest"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cars"
                ],
                "summary": "Release by ID",
                "description": "Remove the specified resource from storage.",
                "operationId": "carFree",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of car",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Car is free"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "CarBusyStore": {
                "title": "Car storring request",
                "description": "Some simple request createa as example",
                "properties": {
                    "user_id": {
                        "title": "User",
                        "description": "User of key for storring",
                        "type": "integer",
                        "example": "4"
                    },
                    "car_id": {
                        "title": "Car",
                        "description": "Value for storring",
                        "type": "integer",
                        "example": "7"
                    }
                },
                "type": "object"
            },
            "CarShowRequest": {
                "title": "Car showing request",
                "description": "Some simple request createa as example",
                "properties": {
                    "user_id": {
                        "title": "User",
                        "description": "Unique ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "car_id": {
                        "title": "Car",
                        "description": "Unique ID",
                        "type": "integer",
                        "example": "3"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "X-APP-ID": {
                "type": "apiKey",
                "name": "X-APP-ID",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Cars",
            "description": "Cars use now"
        }
    ]
}